[Refactor] : Product -> Post로 변경
