+++ src/main/java/com/daangn/dangunmarket/domain/post/facade/mpper/PostParamMapper.java
@@ -1,0 +1,32 @@
+package com.daangn.dangunmarket.domain.post.facade.mpper;
+
+import com.daangn.dangunmarket.domain.post.facade.dto.PostCreateRequestParam;
+import com.daangn.dangunmarket.domain.post.model.Category;
+import com.daangn.dangunmarket.domain.post.model.LocationPreference;
+import com.daangn.dangunmarket.domain.post.service.dto.PostCreateRequest;
+import com.daangn.dangunmarket.domain.post.model.PostImage;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class PostParamMapper {
+
+    public PostCreateRequest toProductCreateRequest(
+            PostCreateRequestParam requestParam,
+            LocationPreference locationPreference,
+            List<PostImage> postImages,
+            Category category) {
+        return new PostCreateRequest(
+                requestParam.memberId(),
+                requestParam.areaId(),
+                locationPreference,
+                postImages,
+                category,
+                requestParam.title(),
+                requestParam.content(),
+                requestParam.price(),
+                requestParam.isOfferAllowed());
+    }
+
+}
