+++ src/main/java/com/daangn/dangunmarket/domain/post/model/Category.java
@@ -1,0 +1,55 @@
+package com.daangn.dangunmarket.domain.post.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.util.Assert;
+
+import java.util.ArrayList;
+import java.util.List;
+
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "name")
+    private String name;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "parent")
+    private Category parent;
+
+    @Column(name = "depth")
+    private Long depth;
+
+    @OneToMany(mappedBy = "parent")
+    private List<Category> children = new ArrayList<>();
+
+    @Builder
+    public Category(String name, Category parent, Long depth, List<Category> children) {
+        Assert.notNull(name, "name은 null값이 들어올 수 없습니다.");
+
+        this.name = name;
+        this.parent = parent;
+        this.depth = depth;
+        this.children = children;
+    }
+
+}
