+++ src/main/java/com/daangn/dangunmarket/domain/post/model/TradeStatus.java
@@ -1,0 +1,6 @@
+package com.daangn.dangunmarket.domain.post.model;
+
+
+public enum TradeStatus {
+    IN_PROGRESS, RESERVATED, FINISH_TRADE
+}
