+++ src/main/java/com/daangn/dangunmarket/domain/post/repository/PostJpaRepository.java
@@ -1,0 +1,7 @@
+package com.daangn.dangunmarket.domain.post.repository;
+
+import com.daangn.dangunmarket.domain.post.model.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostJpaRepository extends JpaRepository<Post, Long> {
+}
