--- src/main/java/com/daangn/dangunmarket/domain/product/controller/dto/ProductCreateApiResponse.java
@@ -1,7 +1,0 @@
-package com.daangn.dangunmarket.domain.product.controller.dto;
-
-public record ProductCreateApiResponse(Long memberId) {
-    public static ProductCreateApiResponse from(Long memberId) {
-        return new ProductCreateApiResponse(memberId);
-    }
-}
