--- src/main/java/com/daangn/dangunmarket/domain/product/model/TradeStatus.java
@@ -1,6 +1,0 @@
-package com.daangn.dangunmarket.domain.product.model;
-
-
-public enum TradeStatus {
-    IN_PROGRESS, RESERVATED, FINISH_TRADE
-}
