--- src/main/java/com/daangn/dangunmarket/domain/product/model/WishList.java
@@ -1,32 +1,0 @@
-package com.daangn.dangunmarket.domain.product.model;
-
-import com.daangn.dangunmarket.global.entity.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "wish_lists")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class WishList extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "products_id", referencedColumnName = "id", nullable = false)
-    private Product product;
-
-    @Column(nullable = false)
-    private Long memberId;
-
-}
