--- src/main/java/com/daangn/dangunmarket/domain/product/repository/ProductRepository.java
@@ -1,12 +1,0 @@
-package com.daangn.dangunmarket.domain.product.repository;
-
-import com.daangn.dangunmarket.domain.product.model.Product;
-
-import java.util.Optional;
-
-public interface ProductRepository {
-
-    Product save(Product product);
-
-    Optional<Product> findById(Long productId);
-}
