--- src/main/java/com/daangn/dangunmarket/domain/product/repository/ProductRepositoryImpl.java
@@ -1,27 +1,0 @@
-package com.daangn.dangunmarket.domain.product.repository;
-
-import com.daangn.dangunmarket.domain.product.model.Product;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public class ProductRepositoryImpl implements ProductRepository {
-
-    private ProductJpaRepository productJpaRepository;
-
-    public ProductRepositoryImpl(ProductJpaRepository productJpaRepository) {
-        this.productJpaRepository = productJpaRepository;
-    }
-
-    @Override
-    public Product save(Product product) {
-        return productJpaRepository.save(product);
-    }
-
-    @Override
-    public Optional<Product> findById(Long productId) {
-        return productJpaRepository.findById(productId);
-    }
-
-}
